pyproject.toml
[project]
name = "claude-usage-monitor"
version = "1.0.0"
description = "Windows system tray application that monitors Claude.ai usage limits"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "winotify>=1.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.3",
    "pyinstaller>=6.0.0",
]

[project.scripts]
claude-monitor = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["src"]
